% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Find the shortest distance}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A dataframe.}

\item{init_node}{the initial node.}
}
\value{
The shortest distance to each node starting from the initial node.
}
\description{
Find the shortest distance
}
\examples{
dijkstra(wiki_graph,1)
dijkstra(wiki_graph,3)
}
