% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Find the greatest common divisor.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{an integer.}

\item{b}{an integer.}
}
\value{
The GCD of \code{a} and \code{b}.
}
\description{
Find the greatest common divisor.
}
\examples{
euclidean(123612,13892347912)
euclidean(100,1000)
}
