% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_doc.R
\docType{package}
\name{shoretespath}
\alias{shoretespath}
\alias{shoretespath-package}
\title{Lab3: A package consisting of Euclidean and Dijkstra algorithms.}
\description{
Lab3: A package consisting of Euclidean and Dijkstra algorithms.
}
\section{The package contains two functions}{
 
\code{\link{euclidean}} that find the greatest common divisor of two integers and
\code{\link{dijkstra}} that finds the shortest distance
from an inital node to all oher nodes in a graph.
}

