% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki_graph}
\format{A data frame with 3 variables:
\describe{
  \item{v1}{nodes}
  \item{v2}{nodes}
  \item{w}{weight of an edge}
  ...
}}
\source{
\url{https://sv.wikipedia.org/wiki/Dijkstras_algoritm}
}
\usage{
wiki_graph
}
\description{
A dataset containing the wiki_graph.
}
\keyword{datasets}
